Move bitwise functions implementations back to core library
Issue #20
